refactor(routine-form): add transition
