+++ src/components/Routine/Drawer/Base/submit-answer-button.tsx
@@ -1,72 +1,26 @@
-import { Box, Button, Flex } from '@chakra-ui/react'
-import styled from '@emotion/styled'
-import React, { useEffect, useRef } from 'react'
+import { Button, Flex } from '@chakra-ui/react'
+import React from 'react'
 import { useIntl } from 'react-intl'
-import { useRecoilState, useRecoilValue } from 'recoil'
+import { useRecoilValue } from 'recoil'
 
-import { OpenArrowDown, OpenArrowUp } from 'src/components/Icon'
 import { currentRoutinePropertiesAtom } from 'src/state/recoil/routine/current-routine-properties'
 import { retrospectiveRoutineIndexQuestionAtom } from 'src/state/recoil/routine/retrospective-showed-question'
 
-import { useRoutineFormAnswers } from '../../hooks/setRoutineFormAnswers/set-routine-form-answers'
-
 import messages from './messages'
 
-const StyledButton = styled(Button)`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 38px;
-  height: 32px;
-  border-radius: 4px;
-  background-color: #6f6eff;
-`
+interface SubmitAnswerButtonProperties {
+  handleClick: () => void
+}
 
-const SubmitAnswerButton = () => {
+const SubmitAnswerButton = ({ handleClick }: SubmitAnswerButtonProperties) => {
   const intl = useIntl()
 
-  const [currentQuestionIndex, setShowedQuestion] = useRecoilState(
-    retrospectiveRoutineIndexQuestionAtom,
-  )
-  const { setRoutineFormAnswers } = useRoutineFormAnswers()
-
-  const buttonReference = useRef<HTMLButtonElement>(null)
-
-  const afterQuestion = () => {
-    setShowedQuestion((currentValue) => currentValue + 1)
-  }
-
+  const currentQuestionIndex = useRecoilValue(retrospectiveRoutineIndexQuestionAtom)
   const { size } = useRecoilValue(currentRoutinePropertiesAtom)
-
-  const handleClick = () => {
-    if (size && currentQuestionIndex < size - 1) afterQuestion()
-    if (size && currentQuestionIndex === size - 1) setRoutineFormAnswers()
-  }
-
-  const comeBack = () => {
-    setShowedQuestion((currentQuestionIndex) => currentQuestionIndex - 1)
-  }
-
-  const handleKeyDown = (event: any) => {
-    const keyCode = event.which || event.key
-
-    if (keyCode === 13 && !event.shiftKey) {
-      event.preventDefault()
-      handleClick()
-    }
-  }
-
-  useEffect(() => {
-    document.addEventListener('keydown', handleKeyDown)
-
-    return () => document.removeEventListener('keydown', handleKeyDown)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
 
   return (
     <Flex alignItems="center">
       <Button
-        ref={buttonReference}
         p="20px 24px"
         color="black.50"
         bg="brand.500"
@@ -84,25 +38,6 @@ const SubmitAnswerButton = () => {
       <span style={{ marginLeft: '12px', color: '#8491B0', fontSize: 14 }}>
         {intl.formatMessage(messages.submitFormButtonDesc)} <b>Enter &#x23CE; &nbsp;</b>
       </span>
-      {currentQuestionIndex > 0 && (
-        <Box position="absolute" right={12} bottom={12} display="flex" gap={1}>
-          {currentQuestionIndex > 0 && (
-            <StyledButton onClick={comeBack}>
-              <OpenArrowUp desc={intl.formatMessage(messages.previousQuestionButtonFormIconDesc)} />
-            </StyledButton>
-          )}
-
-          {currentQuestionIndex > 0 && (
-            <StyledButton
-              disabled={!(size && currentQuestionIndex < size - 1)}
-              _hover={{}}
-              onClick={handleClick}
-            >
-              <OpenArrowDown desc={intl.formatMessage(messages.afterQuestionButtonFormIconDesc)} />
-            </StyledButton>
-          )}
-        </Box>
-      )}
     </Flex>
   )
 }
