+++ src/components/Routine/Drawer/Questions/RoadBlock/index.tsx
@@ -19,6 +19,7 @@ const RoadBlockQuestion = ({
   hidden,
   answer,
   setAnswer,
+  handleClick,
 }: RoadBlockQuestionProperties) => {
   const intl = useIntl()
   const setShowedQuestion = useSetRecoilState(retrospectiveRoutineIndexQuestionAtom)
@@ -44,7 +45,7 @@ const RoadBlockQuestion = ({
   const group = getRootProps()
 
   return (
-    <BaseQuestionRoutineForm>
+    <BaseQuestionRoutineForm handleClick={handleClick}>
       <Stack gap={14}>
         <Text as="h2" color="new-gray.900" fontSize={21} fontWeight="bold">
           {heading}
