+++ src/components/Routine/Drawer/Questions/base.tsx
@@ -1,18 +1,27 @@
-import { Box, Stack } from '@chakra-ui/react'
-import React from 'react'
+import { Box, Stack, SlideFade } from '@chakra-ui/react'
+import React, { useState, useEffect } from 'react'
 
 import SubmitAnswerButton from '../Base/submit-answer-button'
 
 interface BaseQuestionRoutineFormProperties {
   children: JSX.Element | JSX.Element[]
+  handleClick: () => void
 }
 
-const BaseQuestionRoutineForm = ({ children }: BaseQuestionRoutineFormProperties) => {
+const BaseQuestionRoutineForm = ({ children, handleClick }: BaseQuestionRoutineFormProperties) => {
+  const [isQuestionVisible, setIsQuestionVisible] = useState(false)
+
+  useEffect(() => {
+    setTimeout(() => setIsQuestionVisible(true), 300)
+  }, [])
+
   return (
-    <Stack display="flex" flexDir="column" justifyContent="space-between" height="100%">
-      <Box>{children}</Box>
-      <SubmitAnswerButton />
-    </Stack>
+    <SlideFade in={isQuestionVisible} style={{ height: '100%' }}>
+      <Stack display="flex" flexDir="column" justifyContent="space-between" height="100%">
+        <Box>{children}</Box>
+        <SubmitAnswerButton handleClick={handleClick} />
+      </Stack>
+    </SlideFade>
   )
 }
