+++ src/components/Routine/Drawer/Questions/types.ts
@@ -30,5 +30,6 @@ export interface FormQuestion {
   properties?: ValueRangeProperties
   formQuestionIndex: number
   answer?: string
+  handleClick: () => void
   setAnswer?: (questionId: string, value: string, hidden?: boolean) => void
 }
